[rename🚚] 패키지 구조 분리
- config
- repository
- resolver
- constants
분리
