+++ src/main/java/com/spoteditor/backend/config/redis/config/RedisConfiguration.java
@@ -1,5 +1,6 @@
-package com.spoteditor.backend.config.redis;
+package com.spoteditor.backend.config.redis.config;
 
+import com.spoteditor.backend.config.redis.RedisProperties;
 import lombok.RequiredArgsConstructor;
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
@@ -9,7 +10,6 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.JdkSerializationRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
