+++ src/main/java/com/spoteditor/backend/config/security/SecurityConfig.java
@@ -1,10 +1,10 @@
 package com.spoteditor.backend.config.security;
 
-import com.spoteditor.backend.config.jwt.JwtFilter;
-import com.spoteditor.backend.config.jwt.JwtUtils;
+import com.spoteditor.backend.config.jwt.filter.JwtFilter;
+import com.spoteditor.backend.config.jwt.utils.JwtUtils;
 import com.spoteditor.backend.config.oauth.handler.OauthFailureHandler;
 import com.spoteditor.backend.config.oauth.handler.OauthSuccessHandler;
-import com.spoteditor.backend.config.oauth.service.CustomOAuth2AuthorizationRequestResolver;
+import com.spoteditor.backend.config.oauth.resolver.CustomOAuth2AuthorizationRequestResolver;
 import com.spoteditor.backend.config.oauth.service.CustomOauthUserService;
 import com.spoteditor.backend.config.util.CookieUtils;
 import lombok.RequiredArgsConstructor;
