+++ src/main/java/com/spoteditor/backend/user/service/UserTokenService.java
@@ -3,8 +3,8 @@ package com.spoteditor.backend.user.service;
 import com.spoteditor.backend.config.util.CookieUtils;
 import com.spoteditor.backend.global.exception.TokenException;
 import com.spoteditor.backend.user.common.dto.UserIdDto;
-import com.spoteditor.backend.config.jwt.JwtConstants;
-import com.spoteditor.backend.config.jwt.JwtUtils;
+import com.spoteditor.backend.config.jwt.constants.JwtConstants;
+import com.spoteditor.backend.config.jwt.utils.JwtUtils;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.MalformedJwtException;
 import io.jsonwebtoken.security.SignatureException;
