+++ src/test/java/com/spoteditor/backend/config/TestSecurityConfig.java
@@ -1,13 +1,11 @@
 package com.spoteditor.backend.config;
 
-import com.spoteditor.backend.config.jwt.JwtFilter;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.security.web.authentication.preauth.AbstractPreAuthenticatedProcessingFilter;
 
 @TestConfiguration
 public class TestSecurityConfig {
